remove no longer possible acceptance tests and move remaining (#45)
They are no longer possible because `Task#execute()` is no longer
avaiable (since Gradle 5.0). On of the remaining tests was a duplicate
with CpdTest and the other two are moved to CpdPluginTest, such that
CpdAcceptanceTest could be removed.

The removed CpdAcceptanceTests will reappear inside the integTest
sourceSet with one of the next commits. But they need a bigger
refactoring so this might take some time ...